obtener la coleccion de todos los usuarios para poder redigirilos a la vista que le corresponde
