--- my-app/src/components/Navbar.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-// import { NavLink } from 'react-router-dom';
-
-const NavBar = () => {
-  return (
-    <div>
-        {/* <NavLink to='/Menú'>Menú</NavLink> */}
-    </div>
-  );
-};
-
-export default NavBar;
