+++ my-app/src/components/WaiterView.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const WaiterView = () => {
+  return (
+    <div>WaiterView</div>
+  );
+};
+
+export default WaiterView;
