+++ my-app/src/context/UserProvider.jsx
@@ -1,10 +1,12 @@
 import React, { createContext, useState, useContext, useEffect } from 'react';
 import { createUserWithEmailAndPassword, signInWithEmailAndPassword, onAuthStateChanged  } from 'firebase/auth';
-import { auth } from '../firebase/firebase';
+import {collection, onSnapshot, query} from 'firebase/firestore';
+import { auth, db } from '../firebase/firebase';
+// import { getDefaultNormalizer } from '@testing-library/react';
 
 
 export const UserContext = createContext(); 
-export const useAuth = () => {
+export const useAuth = () => {  
     const context = useContext(UserContext);
     if (!context) throw new Error('there is not auth provider');
     return context;
@@ -14,6 +16,7 @@ export const useAuth = () => {
 //children es el resto de los componentes, esto es para que se sigan mostrando los componentes
 const UserProvider = ({ children }) => {
     const [user, setUser] = useState(false);
+    const [userRol, setUserRol] = useState('');
 
     const signup = async (email, password) => {
         await createUserWithEmailAndPassword(auth, email, password);
@@ -23,12 +26,52 @@ const UserProvider = ({ children }) => {
         signInWithEmailAndPassword(auth, email, password);
       };
 
+  const getUserWithRol = () => {
+    const userRolData = query(collection(db, 'users'));
+    console.log(userRolData);
+    let userRolArray = [];
+    onSnapshot(userRolData, (querySnapshot) => {
+      querySnapshot.forEach(doc => {
+        userRolArray.push(
+        {
+          id: doc.id,
+          doc: doc.data()
+        }
+        );
+      });
+       return setUserRol(userRolArray);
+     });
+    };
+   
+      
+    
+      // function setUserWithFirebaseAndRol(currentUser) {
+      //   getUserWithRol(currentUser).forEach(rol =>{
+          
+      //     {
+      //       currentUser.uid,
+      //       currentUser.email,
+      //       currentUser.password,
+      //       rol,
+      //     }
+
+      //   });
+      //     setUserRol(currentUser);
+      //     console.log('userData final', userData);
+      //   // });
+      // }
+
       useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
           if (currentUser) {
-            // const {email, displayName, uid} = user
+            // const {email, displayName, uid} = user}
             setUser(currentUser);
-          } else {
+            getUserWithRol(currentUser);
+           }
+          // if (!user){
+          //   setUserWithFirebaseAndRol(currentUser);
+          // }
+           else {
             setUser(null);
           }
         });
@@ -38,7 +81,7 @@ const UserProvider = ({ children }) => {
       }, []);
 
     return (
-        <UserContext.Provider value={{ user, setUser, signup, login }}>
+        <UserContext.Provider value={{signup, login, user, userRol }}>
         {children}
         </UserContext.Provider>
     );
